requirements.txt
CherryPy==18.10.0
Jinja2==3.1.6
python-dotenv==1.0.1
python-telegram-bot==22.0
ldap3==2.9.1